+++ srchybrid/DropDownButton.cpp
@@ -1,5 +1,5 @@
 //this file is part of eMule
-//Copyright (C)2005 Merkur ( devs@emule-project.net / http://www.emule-project.net )
+//Copyright (C)2002-2006 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / http://www.emule-project.net )
 //
 //This program is free software; you can redistribute it and/or
 //modify it under the terms of the GNU General Public License
@@ -140,7 +140,7 @@ void CDropDownButton::ResizeToMaxWidth()
 	    TBBUTTONINFO tbbi = {0};
 	    tbbi.cbSize = sizeof tbbi;
 	    tbbi.dwMask = TBIF_SIZE;
-	    tbbi.cx = rcWnd.Width();
+	    tbbi.cx = (WORD)rcWnd.Width();
 	    SetButtonInfo(GetWindowLong(m_hWnd, GWL_ID), &tbbi);
 	}
 }
