+++ srchybrid/KnownFileList.h
@@ -1,5 +1,5 @@
 //this file is part of eMule
-//Copyright (C)2002 Merkur ( devs@emule-project.net / http://www.emule-project.net )
+//Copyright (C)2002-2006 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / http://www.emule-project.net )
 //
 //This program is free software; you can redistribute it and/or
 //modify it under the terms of the GNU General Public License
@@ -35,7 +35,7 @@ public:
 	void	Clear();
 	void	Process();
 
-	CKnownFile* FindKnownFile(LPCTSTR filename, uint32 date, uint32 size) const;
+	CKnownFile* FindKnownFile(LPCTSTR filename, uint32 date, uint64 size) const;
 	CKnownFile* FindKnownFileByID(const uchar* hash) const;
 	CKnownFile* FindKnownFileByPath(const CString& sFilePath) const;
 	void	MergePartFileStats(CKnownFile* original);	// SLUGFILLER: mergeKnown - retrieve part file stats from known file
