+++ srchybrid/StatisticFile.h
@@ -1,5 +1,5 @@
 //this file is part of eMule
-//Copyright (C)2002 Merkur ( devs@emule-project.net / http://www.emule-project.net )
+//Copyright (C)2002-2006 Merkur ( strEmail.Format("%s@%s", "devteam", "emule-project.net") / http://www.emule-project.net )
 //
 //This program is free software; you can redistribute it and/or
 //modify it under the terms of the GNU General Public License
@@ -59,8 +59,8 @@ public:
 	/*
 	void	AddTransferred(uint64 bytes);
 	*/
-	void	AddTransferred(uint32 start, uint32 bytes);
-	void	AddBlockTransferred(uint32 start, uint32 end, uint32 count);
+	void	AddTransferred(uint64 start, uint64 bytes);
+	void	AddBlockTransferred(uint64 start, uint64 end, uint64 count);
 	void	DrawSpreadBar(CDC* dc, RECT* rect, bool bFlat) /*const*/;
 	float	GetSpreadSortValue() /*const*/;
 	float	GetFullSpreadCount() /*const*/;
@@ -83,7 +83,7 @@ public:
 	//Morph End - Added by AndCycle, Equal Chance For Each File
 private:
 	//MORPH START - Added by IceCream SLUGFILLER: Spreadbars
-	CRBMap<uint32, uint32> spreadlist;
+	CRBMap<uint64, uint64> spreadlist;
 	static CBarShader s_SpreadBar;
 	//MORPH - Added by SiRoB, Reduce SpreadBar CPU consumption
 	bool	InChangedSpreadSortValue;
